fix - Album#aside - новый Fav...invert

+++ app/models/favorite.rb
@@ -3,11 +3,10 @@ class Favorite < ApplicationRecord
   belongs_to :like, polymorphic: true
 
 
-  def self.invert(user_id:, album_id: nil, track_id: nil)
-    return if user_id.blank?
-    return if album_id.blank? && track_id.blank?
+  def self.invert(user_id:, like_id:, like_type:)
+    return if user_id.blank? || like_id.blank? || like_type.blank?
 
-    test = {user_id: user_id, album_id: album_id, track_id: track_id}
+    test = {user_id: user_id, like_id: like_id, like_type: like_type}
     return delete_by(test) if find_by(test)
 
     create(test)
